// webview-ui/src/thor/model/MediaObject.ts
// tslint:disable
/**
 * ValkyrAI CORE API
 * The API for accessing ValkyrAI services
 *
 * The version of the OpenAPI document: 0.9.25
 * Contact: info@valkyrlabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../src/runtime"

// thorapi

/**
 * GridHeim Embedded Media Object
 * @export
 * @interface MediaObject
 */
export type MediaObject = {
	/**
	 *
	 * @type {string}
	 * @memberof MediaObject
	 */
	sheetId?: string
	/**
	 *
	 * @type {string}
	 * @memberof MediaObject
	 */
	type?: MediaObjectTypeEnum
	/**
	 *
	 * @type {string}
	 * @memberof MediaObject
	 */
	url?: string
	/**
	 *
	 * @type {number}
	 * @memberof MediaObject
	 */
	positionRow?: number
	/**
	 *
	 * @type {number}
	 * @memberof MediaObject
	 */
	positionCol?: number
	/**
	 * Unique identifier for object in the system
	 * @type {string}
	 * @memberof MediaObject
	 */
	id?: string
	/**
	 * UUID of owner of the object in the system
	 * @type {string}
	 * @memberof MediaObject
	 */
	ownerId?: string
	/**
	 * Date of object creation
	 * @type {Date}
	 * @memberof MediaObject
	 */
	createdDate?: Date
	/**
	 * Data, including hash of the key(s) used to encrypt this record.
	 * @type {string}
	 * @memberof MediaObject
	 */
	keyHash?: string
	/**
	 * Last user to access object
	 * @type {string}
	 * @memberof MediaObject
	 */
	lastAccessedById?: string
	/**
	 * Timestamp of last access of object
	 * @type {Date}
	 * @memberof MediaObject
	 */
	lastAccessedDate?: Date
	/**
	 * Unique identifier for user who last modifed the object in the system
	 * @type {string}
	 * @memberof MediaObject
	 */
	lastModifiedById?: string
	/**
	 * Date of last object modification
	 * @type {Date}
	 * @memberof MediaObject
	 */
	lastModifiedDate?: Date
}

export function MediaObjectFromJSON(json: any): MediaObject {
	return {
		sheetId: !exists(json, "sheetId") ? undefined : json["sheetId"],
		type: !exists(json, "type") ? undefined : json["type"],
		url: !exists(json, "url") ? undefined : json["url"],
		positionRow: !exists(json, "position:-row") ? undefined : json["position:-row"],
		positionCol: !exists(json, "position-col") ? undefined : json["position-col"],
		id: !exists(json, "id") ? undefined : json["id"],
		ownerId: !exists(json, "ownerId") ? undefined : json["ownerId"],
		createdDate: !exists(json, "createdDate") ? undefined : new Date(json["createdDate"]),
		keyHash: !exists(json, "keyHash") ? undefined : json["keyHash"],
		lastAccessedById: !exists(json, "lastAccessedById") ? undefined : json["lastAccessedById"],
		lastAccessedDate: !exists(json, "lastAccessedDate") ? undefined : new Date(json["lastAccessedDate"]),
		lastModifiedById: !exists(json, "lastModifiedById") ? undefined : json["lastModifiedById"],
		lastModifiedDate: !exists(json, "lastModifiedDate") ? undefined : new Date(json["lastModifiedDate"]),
	}
}

export function MediaObjectToJSON(value?: MediaObject): any {
	if (value === undefined) {
		return undefined
	}
	return {
		sheetId: value.sheetId,
		type: value.type,
		url: value.url,
		"position:-row": value.positionRow,
		"position-col": value.positionCol,
		id: value.id,
		ownerId: value.ownerId,
		createdDate: value.createdDate === undefined ? undefined : value.createdDate.toISOString(),
		keyHash: value.keyHash,
		lastAccessedById: value.lastAccessedById,
		lastAccessedDate: value.lastAccessedDate === undefined ? undefined : value.lastAccessedDate.toISOString(),
		lastModifiedById: value.lastModifiedById,
		lastModifiedDate: value.lastModifiedDate === undefined ? undefined : value.lastModifiedDate.toISOString(),
	}
}

/**
 * @export
 * @enum {string}
 */
export enum MediaObjectTypeEnum {
	MEDIAIMAGE = "image",
	MEDIAAUDIO = "audio",
	MEDIAVIDEO = "video",
	MEDIASHAPE = "shape",
}